AG-13178 - Rename fibonacci retracement folder
