+++ packages/ag-charts-enterprise/src/features/annotations/annotationsConfig.ts
@@ -4,7 +4,7 @@ import { calloutConfig } from './callout/calloutConfig';
 import { commentConfig } from './comment/commentConfig';
 import { horizontalLineConfig, verticalLineConfig } from './cross-line/crossLineConfig';
 import { disjointChannelConfig } from './disjoint-channel/disjointChannelConfig';
-import { fibonacciRetracementConfig } from './fibonacciRetracement/fibonacciRetracementConfig';
+import { fibonacciRetracementConfig } from './fibonacci-retracement/fibonacciRetracementConfig';
 import { arrowConfig, lineConfig } from './line/lineConfig';
 import {
     datePriceRangeConfig,
